functions/requirements.txt
# Azure Function requirements
azure-functions>=1.18.0
openai>=1.3.0
google-generativeai>=0.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.4.0
